Add leaderboard scene
The leaderboard scene is shown after a set amount of matches (I used 10, but it's adjustable) and displays the top 10 robot names and current records, so that people have a better sense of who is doing well. The robots are sorted by win percentage, and then by number of matches if that is tied.